+++ benchmark/dirty.js
@@ -2,27 +2,32 @@
 const planckmatch = require(`../library`);
 // Get shared functions.
 const { before, beforeEach, afterEach, options } = require(`./_shared`);
+const fileCount = options.fileCount,
+	filePaths = options.filePaths,
+	iterationCount = options.iterationCount,
+	pattern = options.pattern,
+	patterns = options.patterns;
 
 // Run before benchmark.
-before(`The following benchmark only uses the planckmatch(value, patterns) function which is only recommended when using that set of patterns once.`);
+before(`The following benchmark only uses the planckmatch(value, patterns) function which is only recommended when using a set of patterns once.`);
 
 // High resolution time.
 let time;
 
 beforeEach(`  Using the first pattern.`);
 time = process.hrtime();
-for (let i = 0; i < options.iterationCount; i++) {
-	for (let j = 0; j < options.filePaths.length; j++) {
-		planckmatch(options.filePaths[j], options.pattern);
+for (let i = 0; i < iterationCount; i++) {
+	for (let j = 0; j < fileCount; j++) {
+		planckmatch(filePaths[j], pattern);
 	}
 }
 afterEach(process.hrtime(time));
 
 beforeEach(`  Using all patterns.`);
 time = process.hrtime();
-for (let i = 0; i < options.iterationCount; i++) {
-	for (let j = 0; j < options.filePaths.length; j++) {
-		planckmatch(options.filePaths[j], options.patterns);
+for (let i = 0; i < iterationCount; i++) {
+	for (let j = 0; j < fileCount; j++) {
+		planckmatch(filePaths[j], patterns);
 	}
 }
 afterEach(process.hrtime(time));